Extract common user-management code to manage-users loader and action
