+++ test/routes/users/delete-user.test.ts
@@ -1,5 +1,4 @@
 import { fromPartial } from '@total-typescript/shoehorn';
-import type { AuthHelper } from '../../../app/auth/auth-helper.server';
 import { action } from '../../../app/routes/users/delete-user';
 import type { UsersService } from '../../../app/users/UsersService.server';
 
@@ -7,11 +6,9 @@ describe('delete-user', () => {
   describe('action', () => {
     const deleteUser = vi.fn().mockResolvedValue(undefined);
     const usersService: UsersService = fromPartial({ deleteUser });
-    const getSession = vi.fn().mockResolvedValue({ role: 'admin' });
-    const authHelper: AuthHelper = fromPartial({ getSession });
     const runAction = (userId: string) => {
       const request = fromPartial<Request>({ json: vi.fn().mockResolvedValue({ userId }) });
-      return action(fromPartial({ request }), usersService, authHelper);
+      return action(fromPartial({ request }), usersService);
     };
 
     it.each([['123', '456']])('deletes provided user by id', async (userId) => {
